+++ js/main.js
@@ -1,5 +1,8 @@
 import './display-picture.js';
 import './form.js';
 import {validate} from './validation.js';
+import './image-scale.js';
+import {onApplyEffect} from './effects-control.js';
 
+onApplyEffect();
 validate();
